theme: split view functions for deposits and records
