+++ invenio_app_rdm/theme/views/__init__.py
@@ -10,6 +10,7 @@
 from flask import Blueprint
 from flask_menu import current_menu
 
+from .deposits import register_deposits_ui_routes
 from .filters import register_filters
 from .general import register_general_ui_routes
 from .records import file_download_ui, register_records_ui_routes
@@ -27,14 +28,13 @@ def ui_blueprint(app):
     register_filters(blueprint)
     register_general_ui_routes(app, blueprint)
     register_records_ui_routes(app, blueprint)
+    register_deposits_ui_routes(app, blueprint)
 
     @blueprint.before_app_first_request
     def init_menu():
         """Initialize menu before first request."""
         item = current_menu.submenu("main.deposit")
-        item.register(
-            "invenio_app_rdm.deposits_user", "Uploads", order=1
-        )
+        item.register("invenio_app_rdm.deposits_user", "Uploads", order=1)
 
     return blueprint
